# Remove unnecessary generated functions