--- a/src/tools/builtins.jl
+++ b/src/tools/builtins.jl
@@ -1,16 +1,13 @@
-@generated function __new__(T, args...)
-  quote
-    Base.@_inline_meta
-    $(Expr(:new, :T, [:(args[$i]) for i = 1:length(args)]...))
-  end
+macro __new__(T, args...)
+  esc(Expr(:new, T, args...))
 end
 
-@generated function __splatnew__(T, args)
-  quote
-    Base.@_inline_meta
-    $(Expr(:splatnew, :T, :args))
-  end
+macro __splatnew__(T, args)
+  esc(Expr(:new, T, args))
 end
+
+@inline __new__(T, args...) = @__splatnew__(T, args)
+@inline __splatnew__(T, args) = @__splatnew__(T, args)
 
 literal_getindex(x, ::Val{i}) where i = getindex(x, i)
 literal_indexed_iterate(x, ::Val{i}) where i = Base.indexed_iterate(x, i)