--- a/src/tools/builtins.jl
+++ b/src/tools/builtins.jl
@@ -3,7 +3,7 @@ macro __new__(T, args...)
 end
 
 macro __splatnew__(T, args)
-  esc(Expr(:new, T, args))
+  esc(Expr(:splatnew, T, args))
 end
 
 @inline __new__(T, args...) = @__splatnew__(T, args)