I could be wrong here, but I suspect that it's completely unnecessary for `__new__` and `__splatnew__` to be `@generated` functions, and their presence likely has a negative impact on the compiler heuristics as they apply to Zygote, since it has to compile a generated function for every combination of types and arguments. 

I've replaced them with just regular functions that use macros instead. I suspect these functions could even be marked `@nospecialize` since all they're meant to do is forward their arguments to the macro and then julia's builtin `:new` and `:splatnew` can handle the rest.