@mcabbott 2022-05-09
Could you time maybe this example: https://github.com/FluxML/Zygote.jl/issues/1126#issue-1065022370 to see if it helps TTFG?