@MasonProtter 2022-05-09
Nice example
```
  1.470254 seconds (9.08 M allocations: 470.487 MiB, 4.84% gc time, 99.85% compilation time)
  0.000304 seconds (1.07 k allocations: 193.859 KiB)
 45.332908 seconds (57.14 M allocations: 3.022 GiB, 1.41% gc time, 99.98% compilation time)
  0.001425 seconds (4.76 k allocations: 1.236 MiB)
```
after
```
  1.524489 seconds (9.08 M allocations: 470.256 MiB, 6.20% gc time, 99.96% compilation time)
  0.000332 seconds (1.07 k allocations: 193.859 KiB)
 44.100403 seconds (57.04 M allocations: 3.016 GiB, 2.02% gc time, 99.97% compilation time)
  0.001552 seconds (4.75 k allocations: 1.236 MiB)
```
So not much of a win, but I guess a very slight one in total for that example.