@MasonProtter 2022-05-09
`@nospecialize` is unfortunatly quite harmful here. 
```
  1.530241 seconds (9.08 M allocations: 470.255 MiB, 6.22% gc time, 99.96% compilation time)
  0.000342 seconds (1.07 k allocations: 193.766 KiB)
 90.697882 seconds (57.05 M allocations: 3.016 GiB, 0.98% gc time, 99.99% compilation time)
  0.001325 seconds (4.76 k allocations: 1.236 MiB)
```