@ToucheSir 2022-05-09
Given how few constructors there are to augment in that example, a 1s increase ain't bad.

> `@nospecialize` is unfortunatly quite harmful here.

At this point I have no idea whether/where more or less stability helps with Zygote. The compile vs runtime divide makes things quite hazy.