@MasonProtter 2022-05-09
> a 1s increase ain't bad

Do you mean a 1s decrease?