@MasonProtter 2022-05-09
Wait, sorry I messed up when I did the `Before` timing! This is actually very profitable. Here is the actual before:
```
  1.525220 seconds (9.08 M allocations: 470.257 MiB, 6.05% gc time, 99.96% compilation time)
  0.001711 seconds (1.68 k allocations: 234.203 KiB, 80.54% compilation time)
 82.328894 seconds (57.15 M allocations: 3.022 GiB, 1.01% gc time, 99.99% compilation time)
  0.001116 seconds (4.77 k allocations: 1.237 MiB)
```
I'll update my above comment.