@mcabbott 2022-05-09
That looks great!

Would be worth `@btime`-ing the final gradient, since 0.001552 / 0.001116 is a 40% slowdown.