@MasonProtter 2022-05-09
Good point. Here's what I'm seeing with BenchmarkTools.jl

Before:
```
julia> @benchmark loss_grad($lr_images, $ps)
BenchmarkTools.Trial: 5568 samples with 1 evaluation.
 Range (min … max):  745.229 μs …   6.328 ms  ┊ GC (min … max): 0.00% … 78.50%
 Time  (median):     801.364 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   895.402 μs ± 525.878 μs  ┊ GC (mean ± σ):  7.18% ± 10.17%

  █▅▄▄▁▂▁                                                       ▁
  ███████▅▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▆█ █
  745 μs        Histogram: log(frequency) by time       4.83 ms <

 Memory estimate: 1.24 MiB, allocs estimate: 4726.
```
After:
```
julia> @benchmark loss_grad($lr_images, $ps)
BenchmarkTools.Trial: 5441 samples with 1 evaluation.
 Range (min … max):  766.419 μs …   5.222 ms  ┊ GC (min … max): 0.00% … 78.27%
 Time  (median):     825.599 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   916.375 μs ± 476.182 μs  ┊ GC (mean ± σ):  6.26% ±  9.82%

  █                                                              
  ██▃▅▄▂▂▁▁▁▁▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂▂▂ ▂
  766 μs           Histogram: frequency by time         4.49 ms <

 Memory estimate: 1.24 MiB, allocs estimate: 4728.
```
So not a 40% loss, but also does appear to be a loss.