@MasonProtter 2022-05-10
Hm, this benchmark seems to be quite noisy. E.g. running it again later after walking away from my computer, the differences end up running the other way. 

Before
```
julia> @benchmark loss_grad($lr_images, $ps)
BenchmarkTools.Trial: 4905 samples with 1 evaluation.
 Range (min … max):  783.859 μs …   6.560 ms  ┊ GC (min … max): 0.00% … 77.99%
 Time  (median):     959.899 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):     1.016 ms ± 503.894 μs  ┊ GC (mean ± σ):  6.09% ±  9.82%

    █▁                                                           
  ▅▃██▃▂▂▂▁▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂▁▂▂▂▂ ▂
  784 μs           Histogram: frequency by time         4.79 ms <

 Memory estimate: 1.24 MiB, allocs estimate: 4728.

```
After
```
julia> @benchmark loss_grad($lr_images, $ps)
BenchmarkTools.Trial: 5892 samples with 1 evaluation.
 Range (min … max):  736.019 μs …   4.558 ms  ┊ GC (min … max): 0.00% … 76.14%
 Time  (median):     782.809 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   846.314 μs ± 411.713 μs  ┊ GC (mean ± σ):  5.96% ±  9.77%

  █▇▁▄▂                                                         ▁
  ██████▅▄▄▃▁▁▃▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▆█ █
  736 μs        Histogram: log(frequency) by time       3.96 ms <

 Memory estimate: 1.24 MiB, allocs estimate: 4734.
```