@mcabbott 2022-05-10
Let's do it. 

I see no change in the TTFG for simple functions, which I guess lies elsewhere. That's about 10s.