@ToucheSir 2022-05-10
10ish seconds is the current fixed overhead of Zygote, so that makes sense to me.